pyproject.toml
[project]
name = "treehacks-amogus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dataset>=1.6.2",
    "datasets>=4.5.0",
    "flask>=3.1.2",
    "langchain>=1.2.10",
    "langgraph>=1.0.8",
    "langgraph-prebuilt>=1.0.7",
    "networkx>=3.6.1",
    "numpy>=2.4.2",
    "openai>=2.21.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tqdm>=4.67.3",
]
